division du Main et des fonctions
